split interpose_vnc.cpp out
